Add files via upload
The first test in this file currently doesn't execute correctly when you run the script in RStudio.  However, the test does run correctly when you select (highlight) the first test and click the Run button.  This appears to be a bug in R or Rstudio, and we're investigating that.  For now, if you want to run these tests, run them by hand (as I describe here).